rename cache folder
